refactor - get user's auth token moved to AuthToken
