+++ YesVancouverApp/src/components/Login/LoginForm.js
@@ -1,8 +1,9 @@
 
 import React, { Component } from 'react'
 import { Alert, StyleSheet, Text, TextInput, TouchableOpacity, View } from 'react-native'
-import {getIndividualContactsList, authenticateContactLogin, getContactEventRegistrationList,
+import {getIndividualContactsList, getContactEventRegistrationList,
     retrieveCurrentContactDetails} from '../Profile/FetchUserDetails'
+import { authenticateContactLogin } from '../../apicalls/Authentication/AuthToken'
 
 let clientEmail = "";
 let clientPassword = "";
@@ -16,10 +17,13 @@ export default class LoginForm extends Component {
     async authenticateLogin(){
         let {navigate} = this.props.navigation;
         let contactAuthenticationToken = await authenticateContactLogin(clientEmail, clientPassword);
+        console.log(contactAuthenticationToken)
         if (contactAuthenticationToken !== null){
             let currentUserDetails = await retrieveCurrentContactDetails(contactAuthenticationToken);
             let contactsListArray = await getIndividualContactsList(currentUserDetails["Id"]);
+            console.log(contactsListArray)
             let contactEventRegistrationDetails = await getContactEventRegistrationList(currentUserDetails["Id"]);
+            console.log(contactEventRegistrationDetails)
             await navigate("NavBar", {'userData' : contactsListArray[0],
                                 'upcomingEvents' : contactEventRegistrationDetails});
         }
